Create Python: Simple Rock Paper, Scissors Game
Rock, Paper, Scissors Game:

This Python program allows a user to play a game of Rock, Paper, Scissors against the computer. The computer randomly selects one of the three options, and the user inputs their choice. The game then determines the winner based on the standard rules:

- Rock beats Scissors
- Scissors beats Paper
- Paper beats Rock

The game provides feedback on the choices and announces the winner.